float

+++ src/test/java/org/broadinstitute/dsp/ontology/performance/results/Stats.java
@@ -60,7 +60,7 @@ public class Stats {
   }
 
   public class OKCounts {
-    long total;
+    float total;
     long ok;
     long ko;
   }
